Use top level Interop functions
The latest API updates allow for the use of the interop functions
without specifying the scope. In this case the coroutines underlying the
execution will be instantiated under the parent's scope. In the case of
these samples, that scope is the scope that is created by the class
implementing CoroutineScope.